Forgotten Realms v1.0 Release Beta
Forgotten Realms is a Metroidvania game based off of the popular game Hollow Knight. The goal of the game is to help save the village from the dragon who lives on a sky island above the village. You, the player, explore the world from a starting hub which is the village to gain more abilities to help defeat the dragon. From exploring deep caves to haunted towers, Forgotten Realms is a game of exploration of combat and exploring. And some things to note are that all the Art is hand-drawn.